hw added class check to service and implemented the required methods

+++ src/main/java/telran/java48/person/dto/CityPopulationDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class CityPopulationDto {
 	String cityString;
-	Integer population;
+	Long population;
 
 }
